로그인페이지와 병합후 구조 수정
