+++ src/App.jsx
@@ -3,19 +3,20 @@ import HomePage from "./pages/HomePage"; // 1. Main 대신 HomePage를 불러옵
 import Character from "./pages/Character";
 import Stocks from "./pages/Stocks";
 import News from "./pages/News";
-import Login from "./component/Login";
-import Register from "./components/page/Register";
-import MyPage from "./components/page/Mypage";
-import Main from "./components/page/Main";
+import Login from "./pages/Login";
+import Register from "./pages/Register";
+import MyPage from "./pages/Mypage";
+import Main from "./pages/Main";
 import Header from "./components/Header";
 import "./index.css"; // 2. Tailwind CSS를 적용하기 위해 CSS 파일을 불러옵니다.
+import Footer from "./components/Footer";
 
 function App() {
   return (
-    <div className="min-h-screen bg-gray-100 flex justify-center items-center font-['Jua']">
-      <div className="bg-slate-950 w-full max-w-md  h-[100vh] flex flex-col">
-        <Header />
-        <div className="overflow-y-auto hide-scrollbar flex flex-col ">
+    <div className="min-h-screen bg-gray-100 flex flex-col items-center font-['Jua']">
+      <Header />
+      <div className="bg-slate-950 w-full max-w-md  flex-1 flex flex-col">
+        <div className="overflow-y-auto hide-scrollbar flex-1 pt-16 ">
           <Routes>
             <Route path="/" element={<Main />} />
             <Route path="/home" element={<HomePage />} />
@@ -28,6 +29,7 @@ function App() {
           </Routes>
         </div>
       </div>
+      <Footer />
     </div>
   );
 }
