+++ src/components/StockListItem.jsx
