+++ src/index.css
@@ -33,6 +33,7 @@
     "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-color: #1e293b;
 }
 
 /* 스크롤바 숨기기 */
